requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.8.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
behave==1.3.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==6.2.0
Deprecated==1.2.18
exceptiongroup==1.3.0
fastapi==0.116.1
frozenlist==1.7.0
future==1.0.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
line-bot-sdk==3.5.0
multidict==6.6.4
oauthlib==3.3.1
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uv==0.8.12
uvicorn==0.35.0
wrapt==1.17.3
yarl==1.20.1
sentry-sdk[fastapi]